Beginning to condense pace_calc and time_conv to one file for ease of use. Trust
