+++ backend/src/utils/pace_calculations.py
@@ -8,6 +8,8 @@ EASY_PCT = 0.67
 TEMPO_PCT = 0.82
 THRESH_PCT = 0.87
 VO2MAX_PCT = 0.95
+TIME_CONVERSION_NUM = 60
+
 
 def get_VDOT(dist: float, time: float) -> float:
     """return the VDOT value of a runner given the distance in meters and time in seconds."""
@@ -29,3 +31,75 @@ def get_training_pace_helper(race_dist: float, race_time: float, pct_pace: float
 
 
 #print(get_training_pace_helper(5000, 17 * 60 + 30, RECOVERY_PCT))
+
+
+
+
+def to_str(sec: int) -> str:
+    """ Convert seconds to a string in the format H:MM:SS."""
+    minutes = (math.floor(sec/TIME_CONVERSION_NUM) %
+               TIME_CONVERSION_NUM)  # Get the number of minutes
+    hours = math.floor(sec/(TIME_CONVERSION_NUM*TIME_CONVERSION_NUM)
+                       )  # Get the number of hours
+    remainingSec = sec % TIME_CONVERSION_NUM  # Get the seconds
+    toReturn = ""
+    if hours != 0:  # Check to add hours
+        toReturn += f"{hours}:"
+    if minutes != 0:  # Check to add minutes
+        toReturn += f"{minutes}:"
+    toReturn += f"{remainingSec:02d}"  # Always add seconds
+    return toReturn
+
+
+def from_str(pace: str) -> int:
+    """ Convert pace in the format H:MM:SS to seconds."""
+    times = pace.split(":")  # Split the string to
+    multiplier = 1  # Prepare the multiplier for conversion
+    total = 0  # Total seconds
+    for num in reversed(times):  # Go through the times to convert to seconds
+        # Add the time to the total when changed by the multiplier
+        total += multiplier * int(num)
+        multiplier *= TIME_CONVERSION_NUM  # Increase the multiplier for the next time
+    return total
+
+
+def from_dec(pace: float) -> str:
+    """ Convert decimal pace to a string in the format M:SS."""
+    dec = math.floor(
+        (pace % 1) * TIME_CONVERSION_NUM)  # Get the decimal part in seconds
+    val = math.floor(pace)  # Get the integer part in minutes
+    return f"{val}:{dec}"
+
+# This is essentially a helper function
+
+
+def total_time_miles(pace, mile: int) -> int:
+    """ Calculate the total time in seconds for a given pace (in seconds per mile) and distance (in miles)."""
+    if isinstance(pace, str):  # If pace is a string, convert it to seconds
+        pace = from_str(pace)
+    return total_time(pace, (mile * MILE))
+
+
+def total_time(pace, distance: int) -> int:
+    """ Calculate the total time in seconds for a given pace (in seconds per mile) and distance (in meters)."""
+    if isinstance(pace, str):  # If pace is a string, convert it to seconds
+        pace = from_str(pace)
+    return math.floor((pace * distance) / MILE)
+
+
+def alter_pace(pace, increase: int) -> int:
+    """ Alter the pace by a given increase in seconds."""
+    if isinstance(pace, str):  # If pace is a string, convert it to seconds
+        pace = from_str(pace)
+    return pace + increase
+
+
+def mile_pace(pace, distance: int) -> int:
+    if distance == 0:  # If distance is 0, return 0 to avoid division by zero
+        return 0
+    if isinstance(pace, str):  # If pace is a string, convert it to seconds
+        pace = from_str(pace)
+    return math.floor((pace * MILE) / distance)
+
+
+# print(parse_pace("5000-10", "17:30 5k runner"))  # Example (to use replace seconds = user.get_pace(int(distance)) with seconds = "17:30")
